+++ addOns/network/src/main/java/org/zaproxy/addon/network/ExtensionNetwork.java
@@ -353,6 +353,17 @@ public class ExtensionNetwork extends ExtensionAdaptor implements CommandLineLis
         return getConnectionOptions().isHttpProxyEnabled();
     }
 
+    /**
+     * Tells whether or not an outbound proxy (HTTP or SOCKS) is enabled.
+     *
+     * @return {@code true} if a proxy is enabled, {@code false} otherwise.
+     * @since 0.30.0
+     */
+    public boolean isProxyEnabled() {
+        return getConnectionOptions().isHttpProxyEnabled()
+                || getConnectionOptions().isSocksProxyEnabled();
+    }
+
     ClientCertificatesOptions getClientCertificatesOptions() {
         return clientCertificatesOptions;
     }